requirements.txt
bs4==0.0.1
begins==0.9
